src/GLC/requirements.txt
confusable-homoglyphs==3.2.0
Django==2.2.1
django-ckeditor==5.7.0
django-crispy-forms==1.7.2
django-db-signals==0.1.1
django-filter==2.1.0
django-js-asset==1.2.2
django-registration==3.0
Pillow==6.0.0
psycopg2==2.8.2
psycopg2-binary==2.8.2
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
